transfer over to Book and Booklist
