+++ src/main/java/seedu/duke/task/Book.java
@@ -1,9 +1,19 @@
 package seedu.duke.task;
 
+import seedu.duke.DukeException;
+import seedu.duke.common.Messages;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
 /**
  * Represents a task in the task list.
  */
 public class Book extends Item {
+
+    protected boolean isReturn;
+    protected LocalDate futureDate;
 
     /**
      * Constructor used when adding a new task.
@@ -25,5 +35,73 @@ public class Book extends Item {
      */
     public Book(String description, boolean isReturn) {
         super(description, isReturn);
+    }
+
+    /**
+     * Marks the task as done and book as returned.
+     */
+    public void markAsReturn() {
+        isReturn = true;
+    }
+
+    /**
+     * Retrieves whether the task in done.
+     *
+     * @return true if the task is done already, false otherwise
+     */
+    public boolean getIsReturn() {
+        return isReturn;
+    }
+
+    @Override
+    public void setDateFromString(String dateString) throws DukeException {
+        assert dateString != null : "dateString should not be null.";
+        try {
+            date = LocalDate.parse(dateString, DATETIME_PARSE_FORMAT);
+            futureDate = date.plusMonths(1);
+        } catch (DateTimeParseException e) {
+            throw new DukeException(Messages.EXCEPTION_INVALID_DATE);
+        }
+    }
+
+    public String getFutureDateString(DateTimeFormatter formatter) {
+        if (futureDate == null) {
+            return "";
+        }
+
+        return futureDate.format(formatter);
+
+    }
+
+    /**
+     * Converts the attributes of the book into a formatted string to be displayed to the user.
+     *
+     * @return the formatted string to be displayed to the user
+     */
+    public String toStringBook(boolean isList) {
+        String returnString = "";
+        if (isList) {
+            if (this.isReturn) {
+                returnString = "[B][R] " + this.description + "\n";
+            } else {
+                returnString = "[B][L] " + this.description + "\n";
+            }
+        } else {
+            returnString = this.description + "\n";
+        }
+        if (date != null) {
+            returnString += "\t\t (Loan Date: " + getDateString(Task.DATETIME_PRINT_FORMAT) + ")\n";
+            returnString += "\t\t (Due Date: " + getFutureDateString(Task.DATETIME_PRINT_FORMAT) + ")";
+        }
+        return returnString;
+    }
+
+
+    public String toFileBook() {
+        String isDoneString = (isDone) ? "1" : "0";
+        String dateString = getDateString(Item.DATETIME_PARSE_FORMAT);
+        String futureDateString = getFutureDateString(Item.DATETIME_PARSE_FORMAT);
+
+        return "B | " + isDoneString + " | " + description + " | " + dateString + " | " + futureDateString;
     }
 }
