+++ src/main/java/seedu/duke/task/BookList.java
@@ -26,12 +26,44 @@ public class BookList extends ItemList<Book> {
         items = new ArrayList<>();
     }
 
+    /**
+     * Adds a book to the book list from the parameters.
+     *
+     * @param book Book to be added to the book list.
+     */
+    public void addBook(Book book) {
+        items.add(book);
+        Ui.dukePrint(Messages.MESSAGE_ADDBOOK + book.toStringBook(false));
+    }
+
     @Override
     public void addTodo(String description) {
         Book newBook = new Book(description);
         items.add(newBook);
         Ui.dukePrint(Messages.MESSAGE_ADDTASK + newBook.toString() + Messages.MESSAGE_STATUS_FIRST
                 + items.size() + Messages.MESSAGE_STATUS_LAST);
+    }
+
+
+    public void markAsReturn(int index) {
+        if (index > items.size() || index < 1) {
+            Ui.dukePrint(Messages.WARNING_NO_TASK);
+        } else {
+            items.get(index - 1).markAsReturn();
+            Ui.dukePrint(Messages.MESSAGE_RETURNED + items.get(index - 1).getDescription());
+        }
+    }
+
+    public void listBook() {
+        String message = "";
+        if (items.size() == 0) {
+            Ui.dukePrint(Messages.MESSAGE_EMPTY_BOOK_LIST);
+            return;
+        }
+        for (int i = 0; i < items.size(); i++) {
+            message += "\n     " + (i + 1) + "." + items.get(i).toStringBook(true);
+        }
+        Ui.dukePrint(Messages.MESSAGE_BOOK_LIST + message);
     }
 }
